WIP: make luks optional

+++ installer.sh
@@ -7,6 +7,7 @@ USERNAME=user
 USER_FULL_NAME="Debian User"
 USER_PASSWORD=hunter2
 ROOT_PASSWORD=changeme
+DISABLE_LUKS=false
 LUKS_PASSWORD=luke
 ENABLE_TPM=true
 HOSTNAME=debian13
@@ -45,9 +46,11 @@ if [ -z "${DISK}" ]; then
     exit 2
 fi
 
-if [ -z "${LUKS_PASSWORD}" ]; then
-    echo "LUKS_PASSWORD variable is missing" >&2
-    exit 3
+if [ "${DISABLE_LUKS}" != "true" ]; then
+  if [ -z "${LUKS_PASSWORD}" ]; then
+      echo "LUKS_PASSWORD variable is missing" >&2
+      exit 3
+  fi
 fi
 
 if [ -z "${NON_INTERACTIVE}" ]; then
@@ -61,9 +64,9 @@ if [ ! -f efi-part.uuid ]; then
     notify generate uuid for efi partition
     uuidgen > efi-part.uuid || exit 1
 fi
-if [ ! -f luks-part.uuid ]; then
-    notify generate uuid for luks partition
-    uuidgen > luks-part.uuid || exit 1
+if [ ! -f main-part.uuid ]; then
+    notify generate uuid for main partition
+    uuidgen > main-part.uuid || exit 1
 fi
 
 if [ "${ENABLE_SWAP}" == "partition" ]; then
@@ -82,41 +85,51 @@ root_part_type="4f68bce3-e8cd-4db1-96e7-fbcaf984b709"  # X86_64
 system_part_type="C12A7328-F81F-11D2-BA4B-00A0C93EC93B"
 swap_part_type="0657FD6D-A4AB-43C4-84E5-0933C84B4F4F "
 efi_part_uuid=$(cat efi-part.uuid)
-luks_part_uuid=$(cat luks-part.uuid)
+main_part_uuid=$(cat main-part.uuid)
+efi_partition=/dev/disk/by-partuuid/${efi_part_uuid}
+main_partition=/dev/disk/by-partuuid/${main_part_uuid}
 btrfs_uuid=$(cat btrfs.uuid)
 top_level_mount=/mnt/top_level_mount
 target=/target
-luks_device=root
-root_device=/dev/mapper/${luks_device}
-kernel_params="rd.luks.options=tpm2-device=auto rw quiet rootfstype=btrfs rootflags=${FSFLAGS},subvol=@ rd.auto=1 splash"
-efi_partition=/dev/disk/by-partuuid/${efi_part_uuid}
-root_partition=/dev/disk/by-partuuid/${luks_part_uuid}
+kernel_params="rw quiet rootfstype=btrfs rootflags=${FSFLAGS},subvol=@ rd.auto=1 splash"
+if [ "${DISABLE_LUKS}" != "true" ]; then
+  kernel_params="rd.luks.options=tpm2-device=auto ${kernel_params}"
+  luks_device=root
+  root_device=/dev/mapper/${luks_device}
+else
+  root_device=${main_partition}
+fi
 
 if [ "${ENABLE_SWAP}" == "partition" ]; then
 swap_part_uuid=$(cat swap-part.uuid)
 swap_size_blocks=$((${SWAP_SIZE}*2048*1024))
 root_start_blocks=$((2099200+${swap_size_blocks}))
 swap_partition=/dev/disk/by-partuuid/${swap_part_uuid}
-swap_device=swap1
-root_partition_nr=3
+swap_device_name=swap1
+if [ "${DISABLE_LUKS}" != "true" ]; then
+  swap_device=/dev/mapper/${swap_device_name}
+else
+  swap_device=${swap_partition}
+fi
+main_partition_nr=3
 sfdisk_format=$(cat <<EOF
 start=2048, size=2097152, type=${system_part_type}, name="EFI system partition", uuid=${efi_part_uuid}
 start=2099200, size=${swap_size_blocks}, type=${swap_part_type}, name="Swap partition", uuid=${swap_part_uuid}
-start=${root_start_blocks}, size=4096000, type=${root_part_type}, name="Root partition", uuid=${luks_part_uuid}
+start=${root_start_blocks}, size=4096000, type=${root_part_type}, name="Root partition", uuid=${main_part_uuid}
 EOF
 )
 else
-root_partition_nr=2
+main_partition_nr=2
 swap_partition=none
 sfdisk_format=$(cat <<EOF
 start=2048, size=2097152, type=${system_part_type}, name="EFI system partition", uuid=${efi_part_uuid}
-start=2099200, size=4096000, type=${root_part_type}, name="LUKS partition", uuid=${luks_part_uuid}
+start=2099200, size=4096000, type=${root_part_type}, name="Root partition", uuid=${main_part_uuid}
 EOF
 )
 fi
 
 if [ ! -f partitions_created.txt ]; then
-notify create ${root_partition_nr} partitions on ${DISK}
+notify create ${main_partition_nr} partitions on ${DISK}
 sfdisk $DISK <<EOF || exit 1
 label: gpt
 unit: sectors
@@ -126,7 +139,7 @@ ${sfdisk_format}
 EOF
 
 notify resize the root partition on ${DISK} to fill available space
-echo ", +" | sfdisk -N ${root_partition_nr} $DISK || exit 1
+echo ", +" | sfdisk -N ${main_partition_nr} $DISK || exit 1
 
 sfdisk -d $DISK > partitions_created.txt || exit 1
 fi
@@ -140,20 +153,20 @@ function wait_for_file {
     done
 }
 
-wait_for_file ${root_partition}
+wait_for_file ${main_partition}
 if [ "${ENABLE_SWAP}" == "partition" ]; then
     wait_for_file ${swap_partition}
 fi
 
-if [ ! -f $KEYFILE ]; then
+if [ "${DISABLE_LUKS}" != "true" -a ! -f $KEYFILE ]; then
     # TODO do we want to store this file in the installed system?
     notify generate key file for luks
     dd if=/dev/random of=${KEYFILE} bs=512 count=1 || exit 1
-    notify "remove any old luks on ${root_partition} (root)"
-    cryptsetup erase --batch-mode ${root_partition}
-    wait_for_file ${root_partition}
-    wipefs -a ${root_partition} || exit 1
-    wait_for_file ${root_partition}
+    notify "remove any old luks on ${main_partition} (root)"
+    cryptsetup erase --batch-mode ${main_partition}
+    wait_for_file ${main_partition}
+    wipefs -a ${main_partition} || exit 1
+    wait_for_file ${main_partition}
     if [ -e ${swap_partition} ]; then
         notify "remove any old luks on ${swap_partition} (swap)"
         cryptsetup erase --batch-mode ${swap_partition}
@@ -179,12 +192,14 @@ function setup_luks {
   cryptsetup luksUUID "$1" > luks.uuid || exit 1
 }
 
-setup_luks ${root_partition}
-root_uuid=$(cat luks.uuid)
+if [ "${DISABLE_LUKS}" != "true" ]; then
+  setup_luks ${main_partition}
+  root_uuid=$(cat luks.uuid)
 
-if [ ! -e ${root_device} ]; then
-    notify open luks on root
-    cryptsetup luksOpen ${root_partition} ${luks_device} --key-file $KEYFILE || exit 1
+  if [ ! -e ${root_device} ]; then
+      notify open luks on root
+      cryptsetup luksOpen ${main_partition} ${luks_device} --key-file $KEYFILE || exit 1
+  fi
 fi
 
 if [ -e /dev/disk/by-partlabel/BaseImage ]; then
@@ -214,19 +229,22 @@ if [ ! -f vfat_created.txt ]; then
 fi
 
 if [ "${ENABLE_SWAP}" == "partition" ]; then
+  if [ "${DISABLE_LUKS}" != "true" ]; then
     setup_luks ${swap_partition} || exit 1
     swap_uuid=$(cat luks.uuid)
 
-    kernel_params="${kernel_params} rd.luks.name=${swap_uuid}=${swap_device} resume=/dev/mapper/${swap_device}"
+    kernel_params="${kernel_params} rd.luks.name=${swap_uuid}=${swap_device_name}"
 
-    if [ ! -e /dev/mapper/${swap_device} ]; then
+    if [ ! -e ${swap_device} ]; then
         notify open luks swap
-        cryptsetup luksOpen ${swap_partition} ${swap_device} --key-file $KEYFILE || exit 1
+        cryptsetup luksOpen ${swap_partition} ${swap_device_name} --key-file $KEYFILE || exit 1
     fi
+  fi
 
-    notify making swap
-    mkswap /dev/mapper/${swap_device} || exit 1
-    swapon /dev/mapper/${swap_device} || exit 1
+  kernel_params="${kernel_params} resume=${swap_device}"
+  notify making swap
+  mkswap ${swap_device} || exit 1
+  swapon ${swap_device} || exit 1
 fi  # swap as partition
 
 if mountpoint -q "${top_level_mount}" ; then
@@ -268,8 +286,8 @@ if [ "${ENABLE_SWAP}" == "file" ]; then
     notify make swap file at ${target}/swap/swapfile
     btrfs filesystem mkswapfile --size ${SWAP_SIZE}G ${target}/swap/swapfile || exit 1
     swapon ${target}/swap/swapfile || exit 1
-    swapfile_offset=$(btrfs inspect-internal map-swapfile -r ${target}//swap/swapfile)
-    kernel_params="${kernel_params} rd.luks.name=${root_uuid}=${luks_device} resume=${root_device} resume_offset=${swapfile_offset}"
+    swapfile_offset=$(btrfs inspect-internal map-swapfile -r ${target}/swap/swapfile)
+    kernel_params="${kernel_params} resume=${root_device} resume_offset=${swapfile_offset}"
 fi
 
 if [ ! -f ${target}/etc/debian_version ]; then
@@ -317,7 +335,7 @@ EOF
 
 if [ "${ENABLE_SWAP}" == "partition" ]; then
 cat <<EOF >> ${target}/etc/fstab || exit 1
-/dev/mapper/${swap_device} swap swap defaults 0 0
+${swap_device} swap swap defaults 0 0
 EOF
 elif [ "${ENABLE_SWAP}" == "file" ]; then
 cat <<EOF >> ${target}/etc/fstab || exit 1
@@ -432,7 +450,7 @@ systemd-cryptenroll --tpm2-device=list > /tmp/tpm-list.txt || exit 1
 if grep -qs "/dev/tpm" /tmp/tpm-list.txt ; then
       echo tpm available, enrolling
       echo "... on root"
-      systemd-cryptenroll --unlock-key-file=/${KEYFILE} --tpm2-device=auto ${root_partition} --tpm2-pcrs=${TPM_PCRS} || exit 1
+      systemd-cryptenroll --unlock-key-file=/${KEYFILE} --tpm2-device=auto ${main_partition} --tpm2-pcrs=${TPM_PCRS} || exit 1
       if [ -e "${swap_partition}" ]; then
           echo "... on swap"
           systemd-cryptenroll --unlock-key-file=/${KEYFILE} --tpm2-device=auto ${swap_partition} --tpm2-pcrs=${TPM_PCRS} || exit 1
@@ -562,17 +580,19 @@ chroot ${target}/ apt-get autoremove -y
 
 notify umounting all filesystems
 if [ "${ENABLE_SWAP}" == "partition" ]; then
-    swapoff /dev/mapper/${swap_device}
+    swapoff ${swap_device}
 elif [ "${ENABLE_SWAP}" == "file" ]; then
     swapoff ${target}/swap/swapfile
 fi
 umount -R ${target}
 umount -R ${top_level_mount}
 
-notify closing luks
-cryptsetup luksClose ${luks_device}
-if [ "${ENABLE_SWAP}" == "partition" ]; then
-  cryptsetup luksClose /dev/mapper/${swap_device}
+if [ "${DISABLE_LUKS}" != "true" ]; then
+  notify closing luks
+  cryptsetup luksClose ${luks_device}
+  if [ "${ENABLE_SWAP}" == "partition" ]; then
+    cryptsetup luksClose /dev/mapper/${swap_device_name}
+  fi
 fi
 
 notify INSTALLATION FINISHED
